Add --no-clean and remove hard-coded no-clean.

+++ mkami-arch.py
@@ -144,6 +144,8 @@ def parse_args():
                         help='The path to the directory on which to mount the root file system during bootstrap.')
     parser.add_argument('--temp-dir', type=str, default='/var/tmp',
                         help='The directory in which to allocate the temporary directory used for bootstrapping.')
+    parser.add_argument('--no-clean', type='store_true', default=False,
+                        help='Do not clean up temporary directory at end, but log path to it (useful for debugging)')
 
     return parser.parse_args()
 
@@ -360,12 +362,14 @@ def main():
             except:
                 log.error('failed to close file like object')
         for tmpdir in tmpdirs:
-            print("WOULD HAVE REMOVED TEMPDIR: {0}".format(tmpdir))
-            continue
-            try:
-                zsh("rm -rf '{0}'")(tmpdir)
-            except:
-                log.error('failed to remove tempdir')
+            if args.no_clean:
+                log.warn('not cleaning %s due to --no-clean', tmpdir)
+                continue
+            else:
+                try:
+                    zsh("rm -rf '{0}'")(tmpdir)
+                except:
+                    log.error('failed to remove tempdir')
 
 if __name__ == '__main__':
     main()
